PreparationInAdv to config.py
