+++ sys.py/UI/constants.py
@@ -12,12 +12,10 @@ from pygame.locals import *
 #from beeprint import pp
 
 #UI lib
-from skin_manager import MySkinManager
 
 
 Width = 320
 Height = 240
-bg_color =  MySkinManager.GiveColor('White')
 
 icon_width  = 80
 icon_height = 80
